requirements.txt
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
pillow==11.1.0
python-pptx==1.0.2
moviepy==1.0.3
pydub==0.25.1
pyttsx3==2.90
pywin32>=306; platform_system=="Windows"
streamlit>=1.38
